add DownloadTaskHelper in order to use more flexible solutions
