+++ app/src/main/java/org/stepic/droid/persistence/downloads/interactor/SectionDownloadInteractor.kt
@@ -1,34 +1,19 @@
 package org.stepic.droid.persistence.downloads.interactor
 
-import org.stepic.droid.persistence.content.StepContentResolver
-import org.stepic.droid.persistence.downloads.DownloadTaskManager
-import org.stepic.droid.persistence.downloads.resolvers.DownloadTitleResolver
 import org.stepic.droid.persistence.downloads.resolvers.structure.StructureResolver
-import org.stepic.droid.persistence.storage.PersistentStateManager
 import org.stepic.droid.persistence.storage.dao.PersistentItemDao
 import org.stepic.droid.persistence.storage.structure.DBStructurePersistentItem
-import org.stepic.droid.storage.repositories.Repository
 import org.stepik.android.model.Section
-import org.stepik.android.model.Step
 import javax.inject.Inject
 
 class SectionDownloadInteractor
 @Inject
 constructor(
         structureResolver: StructureResolver<Section>,
-        downloadTaskManager: DownloadTaskManager,
         persistentItemDao: PersistentItemDao,
 
-        persistentStateManager: PersistentStateManager,
-
-        downloadTitleResolver: DownloadTitleResolver,
-
-        stepContentResolver: StepContentResolver,
-        stepRepository: Repository<Step>
-): DownloadInteractorBase<Section>(
-        structureResolver, downloadTaskManager, persistentItemDao,
-        persistentStateManager, downloadTitleResolver,
-        stepContentResolver, stepRepository) {
+        downloadTasksHelper: DownloadTaskHelper
+): DownloadInteractorBase<Section>(structureResolver, persistentItemDao, downloadTasksHelper) {
     override val Section.keyFieldValue: Long
         get() = id
     override val keyFieldColumn: String = DBStructurePersistentItem.Columns.SECTION
