+++ app/src/main/java/org/stepic/droid/persistence/downloads/interactor/UnitDownloadInteractor.kt
@@ -1,14 +1,8 @@
 package org.stepic.droid.persistence.downloads.interactor
 
-import org.stepic.droid.persistence.content.StepContentResolver
-import org.stepic.droid.persistence.downloads.DownloadTaskManager
-import org.stepic.droid.persistence.downloads.resolvers.DownloadTitleResolver
 import org.stepic.droid.persistence.downloads.resolvers.structure.StructureResolver
-import org.stepic.droid.persistence.storage.PersistentStateManager
 import org.stepic.droid.persistence.storage.dao.PersistentItemDao
 import org.stepic.droid.persistence.storage.structure.DBStructurePersistentItem
-import org.stepic.droid.storage.repositories.Repository
-import org.stepik.android.model.Step
 import org.stepik.android.model.Unit
 import javax.inject.Inject
 
@@ -16,19 +10,10 @@ class UnitDownloadInteractor
 @Inject
 constructor(
         structureResolver: StructureResolver<Unit>,
-        downloadTaskManager: DownloadTaskManager,
         persistentItemDao: PersistentItemDao,
 
-        persistentStateManager: PersistentStateManager,
-
-        downloadTitleResolver: DownloadTitleResolver,
-
-        stepContentResolver: StepContentResolver,
-        stepRepository: Repository<Step>
-): DownloadInteractorBase<Unit>(
-        structureResolver, downloadTaskManager, persistentItemDao,
-        persistentStateManager, downloadTitleResolver,
-        stepContentResolver, stepRepository) {
+        downloadTasksHelper: DownloadTaskHelper
+): DownloadInteractorBase<Unit>(structureResolver, persistentItemDao, downloadTasksHelper) {
     override val Unit.keyFieldValue: Long
         get() = id
     override val keyFieldColumn: String = DBStructurePersistentItem.Columns.UNIT
